forge fmt

+++ src/interfaces/core/ITrustDomainRegistry.sol
@@ -74,17 +74,13 @@ interface ITrustDomainRegistry {
         uint256 s
     ) external;
 
-    function addQE(
-        QEReport memory qeReport,
-        uint256 platformSerial,
-        uint256 pckSerial,
-        uint256 r,
-        uint256 s
-    ) external returns (uint256 qeId);
+    function addQE(QEReport memory qeReport, uint256 platformSerial, uint256 pckSerial, uint256 r, uint256 s)
+        external
+        returns (uint256 qeId);
 
     function addTD(
         TDQuote memory tdQuote,
-        uint qeId,
+        uint256 qeId,
         uint256 x,
         uint256 y,
         bytes32 authentication_data,
@@ -92,9 +88,9 @@ interface ITrustDomainRegistry {
         uint256 s
     ) external returns (uint256 tdId);
 
-    function getRootKey() external view returns(ECKey memory);
-    
-    function getPlatformCAKey(uint256 serial) external view returns(ECKey memory);
+    function getRootKey() external view returns (ECKey memory);
+
+    function getPlatformCAKey(uint256 serial) external view returns (ECKey memory);
 
     function getPCK(uint256 platformSerial, uint256 pckSerial) external view returns (ECKey memory);
 
@@ -111,7 +107,7 @@ interface ITrustDomainRegistry {
     function getQEAuthority(uint256 qeId) external view returns (uint256 platformSerial, uint256 pckSerial);
 
     function getPCKCounterByPlatformCA(uint256 platformSerial) external view returns (uint256);
-    
+
     function getQECounterByProcessorPCK(uint256 platformSerial, uint256 pckSerial) external view returns (uint256);
 
     function getTDCounterByQE(uint256 qeId) external view returns (uint256);
